requirements.txt
fastapi
uvicorn
openai
python-dotenv
chromadb
tiktoken
numpy
pydantic